v2: flatten api/ subdirectory back to repo root
