+++ Dockerfile
@@ -8,10 +8,8 @@ RUN apt-get update && apt-get install -y \
     pkg-config \
     && rm -rf /var/lib/apt/lists/*
 
-# Copy the crate (lives in api/) into the image and build from there
-COPY api ./api
-
-WORKDIR /solana_verified_program_api/api
+# Copy the entire project directory to the Docker image
+COPY . .
 
 RUN cargo build --release
 
@@ -21,7 +19,7 @@ FROM debian:stable-slim AS api_final
 
 WORKDIR /solana_verified_program_api
 
-COPY --from=api_build /solana_verified_program_api/api/target/release/verified_programs_api .
+COPY --from=api_build /solana_verified_program_api/target/release/verified_programs_api .
 
 COPY --from=api_build /usr/local/cargo/bin/solana-verify /usr/local/bin/solana-verify
