+++ src/api/handlers/verified_programs_status.rs
@@ -1,5 +1,5 @@
+use crate::api::responses::{Status, VerifiedProgramsStatusListResponse};
 use crate::db::DbClient;
-use crate::responses::{Status, VerifiedProgramsStatusListResponse};
 use axum::{extract::State, http::StatusCode, Json};
 use tracing::{error, info};
