+++ src/api/mod.rs
@@ -1,0 +1,5 @@
+//! HTTP surface: routes, handlers, response shapes.
+
+pub mod handlers;
+pub mod responses;
+pub mod routes;
