+++ src/db.rs
@@ -6,8 +6,8 @@
 
 use crate::{
     errors::{ApiError, Result},
-    services::onchain::ProgramOnchainState,
-    validation::Address,
+    onchain::ProgramOnchainState,
+    types::Address,
 };
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
@@ -140,8 +140,8 @@ pub struct NewBuild {
     pub signer: Option<Address>,
 }
 
-impl From<&crate::services::onchain::OtterBuildParams> for NewBuild {
-    fn from(p: &crate::services::onchain::OtterBuildParams) -> Self {
+impl From<&crate::onchain::OtterBuildParams> for NewBuild {
+    fn from(p: &crate::onchain::OtterBuildParams) -> Self {
         NewBuild {
             repository: p.git_url.clone(),
             commit_hash: Some(p.commit.clone()),
@@ -343,7 +343,7 @@ impl DbClient {
     pub async fn get_all_verification_info(
         &self,
         program_id: Address,
-    ) -> Result<Vec<crate::responses::VerificationResponseWithSigner>> {
+    ) -> Result<Vec<crate::api::responses::VerificationResponseWithSigner>> {
         let state = self.get_program_state(&program_id).await?;
         let builds = sqlx::query_as::<_, BuildRow>(
             "SELECT DISTINCT ON (signer) * FROM builds
@@ -354,7 +354,7 @@ impl DbClient {
         .fetch_all(&self.pool)
         .await?;
 
-        use crate::responses::{VerificationResponse, VerificationResponseWithSigner};
+        use crate::api::responses::{VerificationResponse, VerificationResponseWithSigner};
         Ok(builds
             .into_iter()
             .map(|b| VerificationResponseWithSigner {
@@ -381,7 +381,7 @@ impl DbClient {
     pub async fn check_is_verified(&self, program_id: Address) -> Result<String> {
         self.verify_cache
             .try_get_with(program_id, async move {
-                let trusted = crate::services::onchain::trusted_signers();
+                let trusted = crate::onchain::trusted_signers();
                 let row: VerificationRow = sqlx::query_as(
                     "SELECT ps.on_chain_hash, ps.is_frozen, ps.is_closed,
                             b.executable_hash, b.repository, b.commit_hash, b.completed_at
@@ -414,8 +414,8 @@ impl DbClient {
                 } else {
                     "On chain program not verified"
                 };
-                let response = crate::responses::ExtendedStatusResponse {
-                    status: crate::responses::StatusResponse {
+                let response = crate::api::responses::ExtendedStatusResponse {
+                    status: crate::api::responses::StatusResponse {
                         is_verified,
                         message: message.to_string(),
                         on_chain_hash,
@@ -424,7 +424,7 @@ impl DbClient {
                             .repository
                             .as_deref()
                             .map(|r| {
-                                crate::services::misc::build_repository_url(
+                                crate::api::responses::build_repository_url(
                                     r,
                                     row.commit_hash.as_deref(),
                                 )
@@ -581,8 +581,8 @@ impl DbClient {
     /// enforce verified-ness in SQL, so each row maps straight to the response.
     pub async fn get_verification_status_all(
         &self,
-    ) -> Result<Vec<crate::responses::VerifiedProgramStatusResponse>> {
-        let trusted = crate::services::onchain::trusted_signers();
+    ) -> Result<Vec<crate::api::responses::VerifiedProgramStatusResponse>> {
+        let trusted = crate::onchain::trusted_signers();
         let builds = sqlx::query_as::<_, BuildRow>(
             "SELECT DISTINCT ON (b.program_id) b.*
              FROM builds b
