+++ src/main.rs
@@ -1,23 +1,22 @@
 use std::net::SocketAddr;
 
 mod api;
+mod build;
 mod config;
 mod db;
 mod errors;
-mod responses;
-mod services;
+mod onchain;
 mod state;
-mod validation;
+mod sweep;
+mod types;
 
 use crate::state::AppState;
 
-/// Result type for API
+/// Result type for the API.
 pub type Result<T> = std::result::Result<T, errors::ApiError>;
 
 #[tokio::main]
 async fn main() {
-    // Initialize logging. `fmt::init()` alone filters everything when
-    // RUST_LOG is unset; default to INFO so deployments aren't silent.
     tracing_subscriber::fmt()
         .with_env_filter(
             tracing_subscriber::EnvFilter::try_from_default_env()
@@ -50,9 +49,9 @@ async fn main() {
         config.max_reverifies_per_sweep,
     );
 
-    services::background_jobs::spawn(state.clone());
+    sweep::spawn(state.clone());
 
-    let app = api::initialize_router(state);
+    let app = api::routes::initialize_router(state);
     let addr = SocketAddr::from(([0, 0, 0, 0], config.port));
     tracing::info!("Server starting on {}", addr);
