+++ src/sweep.rs
@@ -5,11 +5,11 @@
 //! for upgrades the `/pda` webhook missed.
 
 use crate::{
+    build,
     db::NewBuild,
-    services::onchain::{get_otter_verify_params, snapshot_programs},
-    services::verification,
+    onchain::{get_otter_verify_params, snapshot_programs},
     state::AppState,
-    validation::Address,
+    types::Address,
 };
 use solana_pubkey::Pubkey;
 use std::{str::FromStr, sync::atomic::Ordering, time::Duration};
@@ -43,8 +43,8 @@ pub fn spawn(state: AppState) {
 /// sweep's own liveness timestamp (`AppState::last_sweep_at`) rather than
 /// `program_state.last_checked` -- the latter is bumped by every verify and
 /// webhook write, so it would report a dead sweep as healthy.
-pub fn health(state: &AppState) -> crate::responses::BackgroundJobHealth {
-    use crate::responses::{BackgroundJobHealth, BackgroundJobStatus};
+pub fn health(state: &AppState) -> crate::api::responses::BackgroundJobHealth {
+    use crate::api::responses::{BackgroundJobHealth, BackgroundJobStatus};
     // Unix-seconds of the last completed sweep; `0` = none since startup.
     let last_sweep_time = state.last_sweep_at.load(Ordering::Relaxed);
     if last_sweep_time == 0 {
@@ -154,7 +154,7 @@ async fn reverify_one(
     info!("sweep: re-verifying {} (build {})", program_id, build_id);
     let task_state = state.clone();
     tokio::spawn(async move {
-        verification::execute(build_id, new_build, task_state, None).await;
+        build::execute(build_id, new_build, task_state, None).await;
     });
     Ok(())
 }
